requirements.txt
streamlit==1.33.0
pandas==2.2.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
# Langchain ecosystem for Streamlit Cloud compatibility
langchain==0.1.5
langchain-community==0.0.17
langchain-openai==0.0.5
openai==1.14.3
chromadb==0.4.24
tiktoken==0.6.0
sentence-transformers==2.6.1